+++ src/app.js
@@ -1,13 +1,9 @@
 import React from "react";
-import { BrowserRouter as Router, Route } from "react-router-dom";
 import axios from "axios";
 
-import Menu from "./menu";
-import RiskLevelSelector from "./risk-level-selector";
-import Table from "./table";
-import Chart from "./chart";
-
 import Loader from "./components/Loader";
+import Routes from "./Routes";
+import Header from "./components/Header";
 
 export default class App extends React.Component {
 	state = {
@@ -38,15 +34,9 @@ export default class App extends React.Component {
 		return !loaded ? (
 			<Loader />
 		) : (
-			<Router>
-				<div>
-					<Menu />
-					<RiskLevelSelector onChangeRiskLevel={this.onChangeRiskLevel} />
-					<Route exact path="/" component={() => <Table riskLevel={riskLevel} cones={this.cones} />} />
-					<Route path="/table" component={() => <Table riskLevel={riskLevel} cones={this.cones} />} />
-					<Route path="/chart" component={() => <Chart riskLevel={riskLevel} cones={this.cones} />} />
-				</div>
-			</Router>
+			<Routes riskLevel={riskLevel} cones={this.cones}>
+				<Header onChangeRiskLevel={this.onChangeRiskLevel} />
+			</Routes>
 		);
 	}
 }